# swagger.yaml
---
openapi: 3.0.2
info:
  title: gpu_service
  version: 0.1.0
paths:
  "/scrape":
    get:
      summary: Scrape Cards
      operationId: scrape_cards_scrape_get
      responses:
        '200':
          description: this scapes the latest graphics cards from bestbuy website. WARNING do not use this request unless needed. it can take 1-2 min to fulfill. Use the other endpoints for quicker calls to database.
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CardList"
  "/cards":
    get:
      summary: Get All
      operationId: get_all_cards_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CardList"
  "/cards/{sku_value}":
    get:
      summary: Get Item By Sku
      operationId: get_item_by_sku_cards__sku_value__get
      parameters:
      - required: true
        schema:
          title: Sku Value
        name: sku_value
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Card"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/cards/{available}":
    get:
      summary: Get Item By Available
      operationId: get_item_by_available_cards__available__get
      parameters:
      - required: true
        schema:
          title: Available
        name: available
        in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/CardList"
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
components:
  schemas:
    Card:
      title: Card
      required:
      - sku_value
      - card_name
      - available
      type: object
      properties:
        sku_value:
          title: Sku Value
          type: string
        card_name:
          title: Card Name
          type: string
        available:
          title: Available
          type: string
    CardList:
      title: CardList
      required:
      - cards
      type: object
      properties:
        cards:
          title: Cards
          type: array
          items:
            "$ref": "#/components/schemas/Card"
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            "$ref": "#/components/schemas/ValidationError"
    ValidationError:
      title: ValidationError
      required:
      - loc
      - msg
      - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            type: string
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string